+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Added Mastodon API endpoint for updating media descriptions.
+
 ### Changed
 
 - Change MSRV to 1.62.1.